backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
assemblyai>=0.31
google-generativeai>=0.6
requests>=2.31
pydantic>=2.7
pydantic-settings>=2.2
